// src/components/JsonSchemaForm/type.ts
export type Props = {
  schema: any;
  form: any;
  model: any;
  type?: string;
  errorText?: any;
  showErrors?: boolean;
  onChange?: any;
  value?: any;
  setDefault?: any;
  error?: any;
  onChangeValidate?: any;
  onModelChange?: any;
  errors?: any;
  ignore?: any;
  option?: any;
  className?: any;
  mapper?: any;
  evalContext?: any;
  otherProps?: { multiline?: boolean; rows?: any; maxRows?: any };
};
export type Options = {
  path?: any;
  lookup: { [x: string]: any };
  global?: any;
  required?: any;
};
export type Schema = {
  type: string;
  title: any;
  description: any;
  maxLength: any;
  minLength: any;
  readOnly: any;
  readonly: any;
  minimum: any;
  exclusiveMinimum: any;
  maximum: number;
  exclusiveMaximum: any;
  validationMessage: any;
  enumNames: any;
  enum: any;
  isObject: any;
  items: any;
  properties: { [x: string]: any };
  required: any;
};